+++ stories/02-stateful.js
@@ -23,12 +23,12 @@ class StateContainer extends Component {
     <Fragment>
       <button onClick={this.handleOpen}>Open Modal</button>
       <Modal
+        appId="root"
         closeOnEsc={true}
         closeOnOutsideClick={true}
         onClose={this.handleClose}
         onOpen={this.handleOpen}
         open={this.state.open}
-        showClose={true}
       >
         {this.props.children}
       </Modal>